requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2

# LangChain stack
langchain==0.2.16
langchain-groq==0.1.9
langchain-community==0.2.16
numpy==1.26.4

# Optional: requests for future extension
requests==2.32.3
itsdangerous==2.2.0
jinja2==3.1.4
python-multipart==0.0.9

